fuel-dashboard/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
nicegui==2.9.1
duckdb==1.1.3
google-cloud-storage==2.18.2
plotly==5.24.1
pandas==2.2.3
pyarrow==18.1.0
geopy==2.4.1
pydantic==2.10.4
pydantic-settings==2.7.1
